Updated README so it matches code, simplified starting dev server using poetry
