--- tests/test_serializer.py
@@ -1,1 +1,0 @@
-#TODO: Add tests for serializer
